+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `impl From<&[AccountInfo; *_IX_ACCOUNTS_LEN]> for *Accounts` for easier CPIs
 - `deserialize` method for `*IxData`. Not using `BorshDeserialize` trait due to breaking change in trait def between 0.9 and 0.10
 - `derive(PartialEq)` for all typedefs and `*IxArgs` and `*IxData`
+- `**Account` for anchor accounts newtype that includes the discriminant in borsh serde
 
 ## [0.1.4] - 2023-07-21